// generated/src/app/api-models/hold-release-rule-select.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class HoldReleaseRuleSelect {
  keyword: string;
  seqRuleId: number;
  reasonCode: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
}